// src/app/models/masters-mpc/packagingpresentation.ts
export class Packagingpresentation {
    id: number = 0;
    name: string = "";
    idPackingType: number = 0;
    packingType: string = "";
    createdByUser: string = "";
    createdByUserId: number = 0;
    updatedByUser: string = "";
    active: boolean = false;
    initialSetup: boolean = false;
}